requirements.txt
# Core ML
torch>=2.0.0
torch-geometric>=2.4.0

# Chemistry
rdkit

# Data & Viz
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Dashboard
streamlit>=1.28.0

# Utilities
tqdm>=4.65.0
Pillow>=10.0.0
